% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myttest.R
\name{myttest}
\alias{myttest}
\title{Ttest constructor function}
\usage{
myttest(x, y, alpha, paired = FALSE)
}
\arguments{
\item{x}{vector of sample data}

\item{y}{vector of sample data}

\item{alpha}{level of significance}

\item{paired}{denotes whether or not the samples are "paired"}
}
\value{
a list identifying which test was conducted, a Y/N
conclusion as to whether we should reject the NULL of equality
of means at the alpha level of significance, the summary statistics made
by the appropriate t.test and the data in a list
}
\description{
This constructor investigates the data for evidence
in the population variances and an appropriate t-test
is made to produce a list.
}
\details{
A constructor function makes an object of a particular OOP
class and makes a lot of useful statistical information that
methods operate on to produce whatever particulars we need
}
\examples{
\dontrun{myttest(x,y,alpha=0.05,paired=FALSE)}
}
