% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.Rttest}
\alias{plot.Rttest}
\title{Plot method for Rttest}
\usage{
\method{plot}{Rttest}(x, ...)
}
\arguments{
\item{x}{list created from constructor function}

\item{...}{additional arguments}
}
\value{
boxplots of the data when non-paired data is analyzed and
boxplot of differences in data when paired = TRUE with the confidence
interval for the difference of means plotted inside the boxplot of
differences
}
\description{
Creates boxplots of the data when no-paired and paired data is
analyzed
}
\examples{
\dontrun{plot(ans1)}
}
