% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.Rttest}
\alias{print.Rttest}
\title{Print method for Rttest}
\usage{
\method{print}{Rttest}(x, ...)
}
\arguments{
\item{x}{list created from constructor function}

\item{...}{additional arguments}
}
\value{
a (1-alpha)100 percent confidence interval for mux-muy
and the test type used on the data with appropriate headings
}
\description{
Prints a (1-alpha)100 percent confidence interval for the difference of means with
appropriate headings
}
\examples{
\dontrun{print(ans1)}
}
