% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyboxplots.R
\name{shinyboxplots}
\alias{shinyboxplots}
\title{Shiny App Boxplot}
\usage{
shinyboxplots()
}
\value{
two boxplots displayed side-by-side, for two samples
}
\description{
Creates two dynamic boxplots using shiny where the values of sample size, mean,
and standard deviation (which are entered into the function rnorm and populates
x and y) can be changed using widgets
}
\examples{
\dontrun{shinyboxplots()}
}
